% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_basin_volumes.R
\name{compute_basin_volumes}
\alias{compute_basin_volumes}
\title{compute_basin_volumes}
\usage{
compute_basin_volumes(Y, elevation, middle)
}
\arguments{
\item{middle}{}
}
\value{

}
\description{
compute_basin_volumes
}
