% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\name{file_paths}
\alias{file_paths}
\title{file_paths}
\usage{
file_paths(wanted_files, root_path)
}
\description{
Creates a named list of paths for each file type the user wants to use from TISC
}
\details{
@param file_paths list of file types from TISC the user wants to include
 @param root_path the path to the root directory in which TISC models are located

 @importFrom dplyr %>%

 @examples
 @export
}
