% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_basin_data.R
\name{find_basin_data}
\alias{find_basin_data}
\title{Find basin data}
\usage{
find_basin_data(Y, elevation_vector, n)
}
\arguments{
\item{Y}{description...}

\item{elevation_vector}{description...}

\item{n}{description...}
}
\value{

}
\description{
Write description here...
}
\examples{

}
