% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basin_geometry.R
\name{get_basin_geometry}
\alias{get_basin_geometry}
\title{get_basin_geometry}
\usage{
get_basin_geometry(models_basin_data, model_paths)
}
\arguments{
\item{models_basin_data}{output from wrangle_pfl}

\item{model_paths}{list of paths to the pfl files in the root directory}
}
\description{
get_basin_geometry
}
