% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decompaction_data.R
\name{get_decompaction_data}
\alias{get_decompaction_data}
\title{get_decompaction_data}
\usage{
get_decompaction_data(models_well_tops, file_paths)
}
\arguments{
\item{models_well_tops}{output from get_well_tops}

\item{file_paths}{list of the paths to the files within the root directory}
}
\value{

}
\description{
get_decompaction_data
}
