% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subsidence_data.R
\name{get_subsidence_data}
\alias{get_subsidence_data}
\title{get_subsidence_data}
\usage{
get_subsidence_data(decompaction_data, file_paths, model_names)
}
\arguments{
\item{decompaction_data}{output from get_decompaction_data}

\item{file_paths}{list of the paths to the files within the root directory
@param model_names names of the models being used}
}
\description{
get_subsidence_data
}
