% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basin_plotly.R
\name{make_basin_plotly}
\alias{make_basin_plotly}
\title{make_basin_plotly}
\usage{
make_basin_plotly(models_basin_data)
}
\arguments{
\item{plot_path}{}
}
\value{

}
\description{
make_basin_plotly
}
