% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basin_plots_individual.R
\name{make_basin_plots_individual}
\alias{make_basin_plots_individual}
\title{make_basin_plots_individual}
\usage{
make_basin_plots_individual(models_basin_data, vis_dir)
}
\arguments{
\item{models_basin_data}{output from the models_basin_data function}

\item{vis_dir}{the directory path to where you would like to save your plots, end the path with an /}
}
\value{

}
\description{
make_basin_plots_individual
}
