% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basin_plots_small_multiples.R
\name{make_basin_plots_small_multiples}
\alias{make_basin_plots_small_multiples}
\title{make_basin_plots_small_multiples}
\usage{
make_basin_plots_small_multiples(models_basin_data)
}
\arguments{
\item{plot_path}{}
}
\value{

}
\description{
make_basin_plots_small_multiples
}
