% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/make_basin_plots_small_multiples_with_midpoint.R
\name{make_basin_plots_small_multiples_with_midpoint}
\alias{make_basin_plots_small_multiples_with_midpoint}
\title{make_basin_plots_small_multiples_with_midpoint}
\usage{
make_basin_plots_small_multiples_with_midpoint(models_basin_data, vis_dir)
}
\arguments{
\item{models_basin_data}{output from wrangle_pfl}

\item{vis_dir}{the directory path to where you would like to save your plots, end the path with an /}
}
\description{
make_basin_plots_small_multiples_with_midpoint
}
