% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_names.R
\name{model_names}
\alias{model_names}
\title{model_names}
\usage{
model_names(file_paths)
}
\arguments{
\item{file_paths}{List of paths to the pfl files that contain the models being analyzed.}
}
\description{
model_names
}
