% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_paths.R
\name{model_paths}
\alias{model_paths}
\title{model_paths}
\usage{
model_paths(wanted_files, root_path)
}
\arguments{
\item{wanted_files}{List of the file's the user wants to pass from TISC}

\item{root_path}{The path to the directory where all the models are stored}
}
\description{
Creates a named list of paths for each file type the user wants to use from TISC
}
