% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_basin_geometry.R
\name{models_basin_geometry}
\alias{models_basin_geometry}
\title{models_basin_geometry}
\usage{
models_basin_geometry(models_basin_data, model_paths)
}
\arguments{
\item{models_basin_data}{output from wrangle_pfl}

\item{model_paths}{paths to the model files}
}
\value{

}
\description{
models_basin_geometry
}
