% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_well_location.R
\name{models_well_locaton}
\alias{models_well_locaton}
\title{models_well_locaton}
\usage{
models_well_locaton(models_basin_data)
}
\arguments{
\item{models_basin_data}{output from wrangle_pfl}
}
\description{
Location along the transect proscribed in the .PRFL file from tisc.
}
