% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basin_depth.R
\name{plot_basin_depth}
\alias{plot_basin_depth}
\title{plot_basin_depth}
\usage{
plot_basin_depth(models_basin_geometry, vis_dir)
}
\arguments{
\item{models_basin_geometry}{output from models_basin_geometry}

\item{vis_dir}{directory where you want to place your figures}
}
\description{
plot_basin_depth
}
