% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basin_symmetry_data.R
\name{plot_basin_symmetry_data}
\alias{plot_basin_symmetry_data}
\title{plot_basin_symmetry_data}
\usage{
plot_basin_symmetry_data(models_basin_symmetry_data, vis_dir)
}
\arguments{
\item{models_basin_symmetry_data}{output from models_basin_symmetry_data}

\item{vis_dir}{directory where you would like to place your plots}
}
\description{
plot_basin_symmetry_data
}
