% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geohistory_curves.R
\name{plot_geohistory_curves}
\alias{plot_geohistory_curves}
\title{plot_geohistory_curves}
\usage{
plot_geohistory_curves(subsidence_data, vis_dir)
}
\arguments{
\item{subsidence_data}{output from get_subsidence_data}

\item{vis_dir}{directory where you would like to place your plots}
}
\value{

}
\description{
plot_geohistory_curves
}
