% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TISCViewR.R
\docType{package}
\name{TISCViewR}
\alias{TISCViewR}
\title{TISCViewR}
\description{
Organizes and visualizes TISC input and outputs for easy model to model comparison.
}
\author{
Bryan R Thomas \email{bryanrt.bt@gmail.com}
}
