% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_pfl.R
\name{wrangle_pfl}
\alias{wrangle_pfl}
\title{Wrangle pfl files}
\usage{
wrangle_pfl(file_paths, model_names)
}
\arguments{
\item{file_paths}{description...}

\item{model_names}{description...}
}
\description{
Description here...
}
