% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_basin_symmetry_data.R
\name{compute_basin_symmetry_data}
\alias{compute_basin_symmetry_data}
\title{compute_basin_symmetry_data}
\usage{
compute_basin_symmetry_data(models_basin_data, file_paths)
}
\arguments{
\item{models_basin_data}{outputs withing wrangle_pfl}

\item{file_paths}{path to the root directory with all models}
}
\value{

}
\description{
compute_basin_symmetry_data
}
