% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basin_geometry.R
\name{get_basin_geometry}
\alias{get_basin_geometry}
\title{get_basin_geometry}
\usage{
get_basin_geometry(models_basin_data)
}
\arguments{
\item{models_basin_data}{output from wrangle_pfl}
}
\value{

}
\description{
get_basin_geometry
}
