% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decompaction_data.R
\name{get_decompaction_data}
\alias{get_decompaction_data}
\title{get_decompaction_data}
\usage{
get_decompaction_data(models_well_tops, file_paths)
}
\arguments{
\item{file_paths}{}
}
\value{

}
\description{
get_decompaction_data
}
