% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subsidence_data.R
\name{get_subsidence_data}
\alias{get_subsidence_data}
\title{get_subsidence_data}
\usage{
get_subsidence_data(decompaction_data, file_paths)
}
\arguments{
\item{file_paths}{}
}
\value{

}
\description{
get_subsidence_data
}
