% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_well_tops.R
\name{get_well_tops}
\alias{get_well_tops}
\title{get_well_tops}
\usage{
get_well_tops(models_well_location, file_paths, model_names)
}
\arguments{
\item{models_well_location}{output from models_well_location}

\item{file_paths}{paths to all files within the root directory}

\item{model_names}{names of the models being analyzed}
}
\value{

}
\description{
get_well_tops
}
