% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basin_plotly.R
\name{make_basin_plotly}
\alias{make_basin_plotly}
\title{make_basin_plotly}
\usage{
make_basin_plotly(models_basin_data, vis_dir = NULL)
}
\arguments{
\item{models_basin_data}{output from wrangle_pfl}

\item{vis_dir}{the directory path to where you would like to save your plots, end the path with an /}
}
\description{
make_basin_plotly
}
